polo-system/server/requirements.api.txt
# --- Web / API ---
fastapi==0.114.1
uvicorn[standard]==0.30.6
python-multipart==0.0.9
watchfiles==0.24.0
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.2
pymupdf==1.24.9

# --- Database ---
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# --- Security ---
bcrypt==4.1.2

# --- Inference stack ---
transformers==4.44.2
accelerate==0.33.0
peft==0.11.1
bitsandbytes==0.43.1
sentencepiece==0.1.99
einops==0.8.0
safetensors==0.4.4
regex==2024.7.24
numpy==1.26.4